REF: tslibs benchmarks for Timedelta, Period
