+++ asv_bench/benchmarks/timedelta.py
@@ -1,49 +1,11 @@
-import datetime
+"""
+Timedelta benchmarks with non-tslibs dependencies.  See
+benchmarks.tslibs.timedelta for benchmarks that rely only on tslibs.
+"""
 
 import numpy as np
 
-from pandas import (
-    DataFrame,
-    Series,
-    Timedelta,
-    Timestamp,
-    timedelta_range,
-    to_timedelta,
-)
-
-
-class TimedeltaConstructor:
-    def time_from_int(self):
-        Timedelta(123456789)
-
-    def time_from_unit(self):
-        Timedelta(1, unit="d")
-
-    def time_from_components(self):
-        Timedelta(
-            days=1,
-            hours=2,
-            minutes=3,
-            seconds=4,
-            milliseconds=5,
-            microseconds=6,
-            nanoseconds=7,
-        )
-
-    def time_from_datetime_timedelta(self):
-        Timedelta(datetime.timedelta(days=1, seconds=1))
-
-    def time_from_np_timedelta(self):
-        Timedelta(np.timedelta64(1, "ms"))
-
-    def time_from_string(self):
-        Timedelta("1 days")
-
-    def time_from_iso_format(self):
-        Timedelta("P4DT12H30M5S")
-
-    def time_from_missing(self):
-        Timedelta("nat")
+from pandas import DataFrame, Series, Timestamp, timedelta_range, to_timedelta
 
 
 class ToTimedelta:
@@ -86,24 +48,6 @@ class TimedeltaOps:
 
     def time_add_td_ts(self):
         self.td + self.ts
-
-
-class TimedeltaProperties:
-    def setup_cache(self):
-        td = Timedelta(days=365, minutes=35, seconds=25, milliseconds=35)
-        return td
-
-    def time_timedelta_days(self, td):
-        td.days
-
-    def time_timedelta_seconds(self, td):
-        td.seconds
-
-    def time_timedelta_microseconds(self, td):
-        td.microseconds
-
-    def time_timedelta_nanoseconds(self, td):
-        td.nanoseconds
 
 
 class DatetimeAccessor:
